add checks if various external programs are even installed. did not get everywhere but the framework is better now

+++ test/test_config.py
@@ -12,14 +12,13 @@ class TestConfig(unittest.TestCase):
     def test_load_config(self):
         path = "test/static/ex.yaml"
         presto.config.build(path, "test/static/ex.chk", "test/static/nazarov-elim-solvated.xyz")
+        os.remove("test/static/ex.chk")
 
         path = "test/static/ex2.yaml"
         presto.config.build(path, "test/static/ex2.chk", oldchk="test/static/ex.chk")
-
-        path = "test/static/ex3.yaml"
-        presto.config.build(path, "test/static/ex3.chk")
-
-        os.remove("test/static/ex.chk")
         os.remove("test/static/ex2.chk")
-        os.remove("test/static/ex3.chk")
 
+        if presto.config.HAS_G16:
+            path = "test/static/ex3.yaml"
+            presto.config.build(path, "test/static/ex3.chk")
+            os.remove("test/static/ex3.chk")
